moved libs out to their respective repos, reorganized senders -> processors
