+++ lib/cluster/runners/op.js
@@ -14,7 +14,7 @@ module.exports = function(context) {
     if (config.ops_directory) opPath = config.ops_directory;
 
     var paths = {};
-    var types = ['readers', 'processors', 'senders'];
+    var types = ['readers', 'processors'];
 
     function gatherPaths(filePath, rootDir, filename) {
         paths[filename] = filePath;
