+++ lib/processors/elasticsearch_bulk.js
