+++ lib/processors/stdout.js
